Updated project folder structure
