+++ expedition-planner/src/App.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import ExpeditionPlanner from './components/expedition-planner/expedition-planner'
+
+const App = () => {
+    return (
+        <>
+            <ExpeditionPlanner/>
+        </>
+    )
+}
+
+export default App
