+++ expedition-planner/src/assets/ExpeditionPlanner.png
