+++ expedition-planner/src/assets/Party 1.png
