+++ expedition-planner/src/assets/Party 2.png
