+++ expedition-planner/src/assets/Party 3.png
