+++ expedition-planner/src/assets/PurpleBG.png
