+++ expedition-planner/src/components/expedition-planner/expedition-planner.css
@@ -1,0 +1,3 @@
+.expedition-planner-container {
+    height: 100vh;
+}
