+++ expedition-planner/src/components/party-cards/party-cards.css
@@ -1,0 +1,4 @@
+.party-cards-container {
+    display: flex;
+    height: 57%;
+}
