+++ expedition-planner/src/components/party-cards/party-cards.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import './party-cards.css'
+
+const PartyCards = () => {
+  return (
+    <div className="party-cards-container">
+      partyCards
+    </div>
+  )
+}
+
+export default PartyCards
