+++ expedition-planner/src/components/search/search.css
@@ -1,0 +1,27 @@
+.search-component {
+    display: flex;
+    justify-content: center;
+    column-gap: 10px;
+    height: 5%;
+}
+
+.search-input {
+    width: 20vw;
+    min-width: 200px;
+    height: 30px;
+    font-family: 'Poppins', sans-serif;
+}
+
+.submit-button {
+    font-family: 'Poppins', sans-serif;
+    height: 35px;
+    border-radius: 20px;
+    background-color: #9B6694;
+    border: none;
+    color: white;
+    width: 80px;
+}
+
+.submit-button:hover {
+    background: #9080C2;
+}
