+++ expedition-planner/src/index.css
@@ -1,0 +1,7 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap');
+
+body {
+    background-image: url(../src/assets/PurpleBG.png);
+    background-repeat: no-repeat;
+    margin: 0;
+}
