--- src/components/expedition-planner/expedition-planner.css
@@ -1,3 +1,0 @@
-.expedition-planner-container {
-    height: 100vh;
-}
