--- src/components/footer/footer.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import './footer.css'
-
-const Footer = () => {
-  return (
-    <div className='footer-container'>
-      <p className='footer-text'>Website created and designed by Alsiel and Azuly</p>
-    </div>
-  )
-}
-
-export default Footer
