--- src/components/header/header.css
@@ -1,17 +1,0 @@
-.header-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-end;
-
-    height: 35%;
-}
-
-.dream-logo {
-    width: 40%;
-}
-
-.expedition-planner-logo {
-    width: 45%;
-    margin-bottom: 15px;
-}
