--- src/components/party-cards/party-cards.css
@@ -1,4 +1,0 @@
-.party-cards-container {
-    display: flex;
-    height: 57%;
-}
